storage: move store snapshot logic to store_snapshot.go
There was already a lot of snapshot-specific logic in `store.go`,
and #16954 is going to add even more. This change moves all of this
code to a new `store_snapshot.go` file, which makes the transmission
and reception of snapshots more easy to understand.

Release note: None
