+++ pkg/storage/store.go
@@ -18,7 +18,6 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"math"
 	"runtime"
 	"sync"
@@ -77,10 +76,6 @@ const (
 	// ttlStoreGossip is time-to-live for store-related info.
 	ttlStoreGossip = 2 * time.Minute
 
-	// preemptiveSnapshotRaftGroupID is a bogus ID for which a Raft group is
-	// temporarily created during the application of a preemptive snapshot.
-	preemptiveSnapshotRaftGroupID = math.MaxUint64
-
 	// defaultRaftEntryCacheSize is the default size in bytes for a
 	// store's Raft log entry cache.
 	defaultRaftEntryCacheSize = 1 << 24 // 16M
@@ -97,15 +92,6 @@ const (
 	// if a range lease holder experiences a failure causing a missed
 	// gossip update.
 	systemDataGossipInterval = 1 * time.Minute
-
-	// Messages that provide detail about why a preemptive snapshot was rejected.
-	snapshotStoreTooFullMsg = "store almost out of disk space"
-	snapshotApplySemBusyMsg = "store busy applying snapshots and/or removing replicas"
-	storeDrainingMsg        = "store is draining"
-
-	// IntersectingSnapshotMsg is part of the error message returned from
-	// canApplySnapshotLocked and is exposed here so testing can rely on it.
-	IntersectingSnapshotMsg = "snapshot intersects existing range"
 )
 
 var changeTypeInternalToRaft = map[roachpb.ReplicaChangeType]raftpb.ConfChangeType{
@@ -2954,52 +2940,6 @@ func (s *Store) maybeWaitForPushee(
 	return nil, nil
 }
 
-// reserveSnapshot throttles incoming snapshots. The returned closure is used
-// to cleanup the reservation and release its resources. A nil cleanup function
-// and a non-empty rejectionMessage indicates the reservation was declined.
-func (s *Store) reserveSnapshot(
-	ctx context.Context, header *SnapshotRequest_Header,
-) (_cleanup func(), _rejectionMsg string, _err error) {
-	if header.RangeSize == 0 {
-		// Empty snapshots are exempt from rate limits because they're so cheap to
-		// apply. This vastly speeds up rebalancing any empty ranges created by a
-		// RESTORE or manual SPLIT AT, since it prevents these empty snapshots from
-		// getting stuck behind large snapshots managed by the replicate queue.
-	} else if header.CanDecline {
-		storeDesc, ok := s.cfg.StorePool.getStoreDescriptor(s.StoreID())
-		if ok && (!maxCapacityCheck(storeDesc) || header.RangeSize > storeDesc.Capacity.Available) {
-			return nil, snapshotStoreTooFullMsg, nil
-		}
-		select {
-		case s.snapshotApplySem <- struct{}{}:
-		case <-ctx.Done():
-			return nil, "", ctx.Err()
-		case <-s.stopper.ShouldStop():
-			return nil, "", errors.Errorf("stopped")
-		default:
-			return nil, snapshotApplySemBusyMsg, nil
-		}
-	} else {
-		select {
-		case s.snapshotApplySem <- struct{}{}:
-		case <-ctx.Done():
-			return nil, "", ctx.Err()
-		case <-s.stopper.ShouldStop():
-			return nil, "", errors.Errorf("stopped")
-		}
-	}
-
-	s.metrics.ReservedReplicaCount.Inc(1)
-	s.metrics.Reserved.Inc(header.RangeSize)
-	return func() {
-		s.metrics.ReservedReplicaCount.Dec(1)
-		s.metrics.Reserved.Dec(header.RangeSize)
-		if header.RangeSize != 0 {
-			<-s.snapshotApplySem
-		}
-	}, "", nil
-}
-
 // HandleSnapshot reads an incoming streaming snapshot and applies it if
 // possible.
 func (s *Store) HandleSnapshot(
@@ -3015,82 +2955,7 @@ func (s *Store) HandleSnapshot(
 	}
 
 	ctx := s.AnnotateCtx(stream.Context())
-	cleanup, rejectionMsg, err := s.reserveSnapshot(ctx, header)
-	if err != nil {
-		return err
-	}
-	if cleanup == nil {
-		return stream.Send(&SnapshotResponse{
-			Status:  SnapshotResponse_DECLINED,
-			Message: rejectionMsg,
-		})
-	}
-	defer cleanup()
-
-	sendSnapError := func(err error) error {
-		return stream.Send(&SnapshotResponse{
-			Status:  SnapshotResponse_ERROR,
-			Message: err.Error(),
-		})
-	}
-
-	// Check to see if the snapshot can be applied but don't attempt to add
-	// a placeholder here, because we're not holding the replica's raftMu.
-	// We'll perform this check again later after receiving the rest of the
-	// snapshot data - this is purely an optimization to prevent downloading
-	// a snapshot that we know we won't be able to apply.
-	if _, err := s.canApplySnapshot(ctx, header.State.Desc); err != nil {
-		return sendSnapError(
-			errors.Wrapf(err, "%s,r%d: cannot apply snapshot", s, header.State.Desc.RangeID),
-		)
-	}
-
-	if err := stream.Send(&SnapshotResponse{Status: SnapshotResponse_ACCEPTED}); err != nil {
-		return err
-	}
-	if log.V(2) {
-		log.Infof(ctx, "accepted snapshot reservation for r%d", header.State.Desc.RangeID)
-	}
-
-	var batches [][]byte
-	var logEntries [][]byte
-	for {
-		req, err := stream.Recv()
-		if err != nil {
-			return err
-		}
-		if req.Header != nil {
-			return sendSnapError(errors.New("client error: provided a header mid-stream"))
-		}
-
-		if req.KVBatch != nil {
-			batches = append(batches, req.KVBatch)
-		}
-		if req.LogEntries != nil {
-			logEntries = append(logEntries, req.LogEntries...)
-		}
-		if req.Final {
-			snapUUID, err := uuid.FromBytes(header.RaftMessageRequest.Message.Snapshot.Data)
-			if err != nil {
-				return sendSnapError(errors.Wrap(err, "invalid snapshot"))
-			}
-
-			inSnap := IncomingSnapshot{
-				SnapUUID:   snapUUID,
-				Batches:    batches,
-				LogEntries: logEntries,
-				State:      &header.State,
-				snapType:   snapTypeRaft,
-			}
-			if header.RaftMessageRequest.ToReplica.ReplicaID == 0 {
-				inSnap.snapType = snapTypePreemptive
-			}
-			if err := s.processRaftSnapshotRequest(ctx, &header.RaftMessageRequest, inSnap); err != nil {
-				return sendSnapError(errors.Wrap(err.GoError(), "failed to apply snapshot"))
-			}
-			return stream.Send(&SnapshotResponse{Status: SnapshotResponse_APPLIED})
-		}
-	}
+	return s.receiveSnapshot(ctx, header, stream)
 }
 
 func (s *Store) uncoalesceBeats(
@@ -3581,287 +3446,6 @@ func (s *Store) HandleRaftResponse(ctx context.Context, resp *RaftMessageRespons
 	return nil
 }
 
-// OutgoingSnapshotStream is the minimal interface on a GRPC stream required
-// to send a snapshot over the network.
-type OutgoingSnapshotStream interface {
-	Send(*SnapshotRequest) error
-	Recv() (*SnapshotResponse, error)
-}
-
-// SnapshotStorePool narrows StorePool to make sendSnapshot easier to test.
-type SnapshotStorePool interface {
-	throttle(reason throttleReason, toStoreID roachpb.StoreID)
-}
-
-var rebalanceSnapshotRate = settings.RegisterByteSizeSetting(
-	"kv.snapshot_rebalance.max_rate",
-	"the rate limit (bytes/sec) to use for rebalance snapshots",
-	envutil.EnvOrDefaultBytes("COCKROACH_PREEMPTIVE_SNAPSHOT_RATE", 2<<20),
-)
-var recoverySnapshotRate = settings.RegisterByteSizeSetting(
-	"kv.snapshot_recovery.max_rate",
-	"the rate limit (bytes/sec) to use for recovery snapshots",
-	envutil.EnvOrDefaultBytes("COCKROACH_RAFT_SNAPSHOT_RATE", 8<<20),
-)
-
-func snapshotRateLimit(
-	st *cluster.Settings, priority SnapshotRequest_Priority,
-) (rate.Limit, error) {
-	switch priority {
-	case SnapshotRequest_RECOVERY:
-		return rate.Limit(recoverySnapshotRate.Get(&st.SV)), nil
-	case SnapshotRequest_REBALANCE:
-		return rate.Limit(rebalanceSnapshotRate.Get(&st.SV)), nil
-	default:
-		return 0, errors.Errorf("unknown snapshot priority: %s", priority)
-	}
-}
-
-type errMustRetrySnapshotDueToTruncation struct {
-	index, term uint64
-}
-
-func (e *errMustRetrySnapshotDueToTruncation) Error() string {
-	return fmt.Sprintf(
-		"log truncation during snapshot removed sideloaded SSTable at index %d, term %d",
-		e.index, e.term,
-	)
-}
-
-// sendSnapshot sends an outgoing snapshot via a pre-opened GRPC stream.
-func sendSnapshot(
-	ctx context.Context,
-	st *cluster.Settings,
-	stream OutgoingSnapshotStream,
-	storePool SnapshotStorePool,
-	header SnapshotRequest_Header,
-	snap *OutgoingSnapshot,
-	newBatch func() engine.Batch,
-	sent func(),
-) error {
-	start := timeutil.Now()
-	to := header.RaftMessageRequest.ToReplica
-	if err := stream.Send(&SnapshotRequest{Header: &header}); err != nil {
-		return err
-	}
-	// Wait until we get a response from the server.
-	resp, err := stream.Recv()
-	if err != nil {
-		storePool.throttle(throttleFailed, to.StoreID)
-		return err
-	}
-	switch resp.Status {
-	case SnapshotResponse_DECLINED:
-		if header.CanDecline {
-			storePool.throttle(throttleDeclined, to.StoreID)
-			declinedMsg := "reservation rejected"
-			if len(resp.Message) > 0 {
-				declinedMsg = resp.Message
-			}
-			return errors.Errorf("%s: remote declined snapshot: %s", to, declinedMsg)
-		}
-		storePool.throttle(throttleFailed, to.StoreID)
-		return errors.Errorf("%s: programming error: remote declined required snapshot: %s",
-			to, resp.Message)
-	case SnapshotResponse_ERROR:
-		storePool.throttle(throttleFailed, to.StoreID)
-		return errors.Errorf("%s: remote couldn't accept snapshot with error: %s",
-			to, resp.Message)
-	case SnapshotResponse_ACCEPTED:
-	// This is the response we're expecting. Continue with snapshot sending.
-	default:
-		storePool.throttle(throttleFailed, to.StoreID)
-		return errors.Errorf("%s: server sent an invalid status during negotiation: %s",
-			to, resp.Status)
-	}
-
-	// The size of batches to send. This is the granularity of rate limiting.
-	const batchSize = 256 << 10 // 256 KB
-	targetRate, err := snapshotRateLimit(st, header.Priority)
-	if err != nil {
-		return errors.Wrapf(err, "%s", to)
-	}
-
-	// Convert the bytes/sec rate limit to batches/sec.
-	//
-	// TODO(peter): Using bytes/sec for rate limiting seems more natural but has
-	// practical difficulties. We either need to use a very large burst size
-	// which seems to disable the rate limiting, or call WaitN in smaller than
-	// burst size chunks which caused excessive slowness in testing. Would be
-	// nice to figure this out, but the batches/sec rate limit works for now.
-	limiter := rate.NewLimiter(targetRate/batchSize, 1 /* burst size */)
-
-	// Determine the unreplicated key prefix so we can drop any
-	// unreplicated keys from the snapshot.
-	unreplicatedPrefix := keys.MakeRangeIDUnreplicatedPrefix(header.State.Desc.RangeID)
-	n := 0
-	var b engine.Batch
-	for ; ; snap.Iter.Next() {
-		if ok, err := snap.Iter.Valid(); err != nil {
-			return err
-		} else if !ok {
-			break
-		}
-		key := snap.Iter.Key()
-		value := snap.Iter.Value()
-		if bytes.HasPrefix(key.Key, unreplicatedPrefix) {
-			// This should never happen because we're using a
-			// ReplicaDataIterator with replicatedOnly=true.
-			log.Fatalf(ctx, "found unreplicated rangeID key: %v", key)
-		}
-		n++
-		mvccKey := engine.MVCCKey{
-			Key:       key.Key,
-			Timestamp: key.Timestamp,
-		}
-		if b == nil {
-			b = newBatch()
-		}
-		if err := b.Put(mvccKey, value); err != nil {
-			b.Close()
-			return err
-		}
-
-		if len(b.Repr()) >= batchSize {
-			if err := limiter.WaitN(ctx, 1); err != nil {
-				return err
-			}
-			if err := sendBatch(stream, b); err != nil {
-				return err
-			}
-			b = nil
-			// We no longer need the keys and values in the batch we just sent,
-			// so reset ReplicaDataIterator's allocator and allow its data to
-			// be garbage collected.
-			snap.Iter.ResetAllocator()
-		}
-	}
-	if b != nil {
-		if err := limiter.WaitN(ctx, 1); err != nil {
-			return err
-		}
-		if err := sendBatch(stream, b); err != nil {
-			return err
-		}
-	}
-
-	firstIndex := header.State.TruncatedState.Index + 1
-	endIndex := snap.RaftSnap.Metadata.Index + 1
-	logEntries := make([][]byte, 0, endIndex-firstIndex)
-
-	scanFunc := func(kv roachpb.KeyValue) (bool, error) {
-		bytes, err := kv.Value.GetBytes()
-		if err == nil {
-			logEntries = append(logEntries, bytes)
-		}
-		return false, err
-	}
-
-	rangeID := header.State.Desc.RangeID
-
-	if err := iterateEntries(ctx, snap.EngineSnap, rangeID, firstIndex, endIndex, scanFunc); err != nil {
-		return err
-	}
-
-	// Inline the payloads for all sideloaded proposals.
-	//
-	// TODO(tschottdorf): could also send slim proposals and attach sideloaded
-	// SSTables directly to the snapshot. Probably the better long-term
-	// solution, but let's see if it ever becomes relevant. Snapshots with
-	// inlined proposals are hopefully the exception.
-	{
-		var ent raftpb.Entry
-		for i := range logEntries {
-			if err := protoutil.Unmarshal(logEntries[i], &ent); err != nil {
-				return err
-			}
-			if !sniffSideloadedRaftCommand(ent.Data) {
-				continue
-			}
-			if err := snap.WithSideloaded(func(ss sideloadStorage) error {
-				newEnt, err := maybeInlineSideloadedRaftCommand(
-					ctx, rangeID, ent, ss, snap.RaftEntryCache,
-				)
-				if err != nil {
-					return err
-				}
-				if newEnt != nil {
-					ent = *newEnt
-				}
-				return nil
-			}); err != nil {
-				if errors.Cause(err) == errSideloadedFileNotFound {
-					// We're creating the Raft snapshot based on a snapshot of
-					// the engine, but the Raft log may since have been
-					// truncated and corresponding on-disk sideloaded payloads
-					// unlinked. Luckily, we can just abort this snapshot; the
-					// caller can retry.
-					//
-					// TODO(tschottdorf): check how callers handle this. They
-					// should simply retry. In some scenarios, perhaps this can
-					// happen repeatedly and prevent a snapshot; not sending the
-					// log entries wouldn't help, though, and so we'd really
-					// need to make sure the entries are always here, for
-					// instance by pre-loading them into memory. Or we can make
-					// log truncation less aggressive about removing sideloaded
-					// files, by delaying trailing file deletion for a bit.
-					return &errMustRetrySnapshotDueToTruncation{
-						index: ent.Index,
-						term:  ent.Term,
-					}
-				}
-				return err
-			}
-			// TODO(tschottdorf): it should be possible to reuse `logEntries[i]` here.
-			var err error
-			if logEntries[i], err = protoutil.Marshal(&ent); err != nil {
-				return err
-			}
-		}
-	}
-
-	req := &SnapshotRequest{
-		LogEntries: logEntries,
-		Final:      true,
-	}
-	// Notify the sent callback before the final snapshot request is sent so that
-	// the snapshots generated metric gets incremented before the snapshot is
-	// applied.
-	sent()
-	if err := stream.Send(req); err != nil {
-		return err
-	}
-	log.Infof(ctx, "streamed snapshot to %s: kv pairs: %d, log entries: %d, rate-limit: %s/sec, %0.0fms",
-		to, n, len(logEntries), humanizeutil.IBytes(int64(targetRate)),
-		timeutil.Since(start).Seconds()*1000)
-
-	resp, err = stream.Recv()
-	if err != nil {
-		return errors.Wrapf(err, "%s: remote failed to apply snapshot", to)
-	}
-	// NB: wait for EOF which ensures that all processing on the server side has
-	// completed (such as defers that might be run after the previous message was
-	// received).
-	if unexpectedResp, err := stream.Recv(); err != io.EOF {
-		return errors.Errorf("%s: expected EOF, got resp=%v err=%v", to, unexpectedResp, err)
-	}
-	switch resp.Status {
-	case SnapshotResponse_ERROR:
-		return errors.Errorf("%s: remote failed to apply snapshot for reason %s", to, resp.Message)
-	case SnapshotResponse_APPLIED:
-		return nil
-	default:
-		return errors.Errorf("%s: server sent an invalid status during finalization: %s",
-			to, resp.Status)
-	}
-}
-
-func sendBatch(stream OutgoingSnapshotStream, batch engine.Batch) error {
-	repr := batch.Repr()
-	batch.Close()
-	return stream.Send(&SnapshotRequest{KVBatch: repr})
-}
-
 // enqueueRaftUpdateCheck asynchronously registers the given range ID to be
 // checked for raft updates when the processRaft goroutine is idle.
 func (s *Store) enqueueRaftUpdateCheck(rangeID roachpb.RangeID) {
@@ -4284,72 +3868,6 @@ func (s *Store) tryGetOrCreateReplica(
 	}
 	repl.mu.Unlock()
 	return repl, true, nil
-}
-
-// canApplySnapshot returns (_, nil) if the snapshot can be applied to
-// this store's replica (i.e. the snapshot is not from an older incarnation of
-// the replica) and a placeholder can be added to the replicasByKey map (if
-// necessary). If a placeholder is required, it is returned as the first value.
-func (s *Store) canApplySnapshot(
-	ctx context.Context, rangeDescriptor *roachpb.RangeDescriptor,
-) (*ReplicaPlaceholder, error) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.canApplySnapshotLocked(ctx, rangeDescriptor)
-}
-
-func (s *Store) canApplySnapshotLocked(
-	ctx context.Context, rangeDescriptor *roachpb.RangeDescriptor,
-) (*ReplicaPlaceholder, error) {
-	if v, ok := s.mu.replicas.Load(int64(rangeDescriptor.RangeID)); ok &&
-		(*Replica)(v).IsInitialized() {
-		// We have the range and it's initialized, so let the snapshot through.
-		return nil, nil
-	}
-
-	// We don't have the range (or we have an uninitialized
-	// placeholder). Will we be able to create/initialize it?
-	if exRng, ok := s.mu.replicaPlaceholders[rangeDescriptor.RangeID]; ok {
-		return nil, errors.Errorf("%s: canApplySnapshotLocked: cannot add placeholder, have an existing placeholder %s", s, exRng)
-	}
-
-	if exRange := s.getOverlappingKeyRangeLocked(rangeDescriptor); exRange != nil {
-		// We have a conflicting range, so we must block the snapshot.
-		// When such a conflict exists, it will be resolved by one range
-		// either being split or garbage collected.
-		exReplica, err := s.GetReplica(exRange.Desc().RangeID)
-		msg := IntersectingSnapshotMsg
-		if err != nil {
-			log.Warning(ctx, errors.Wrapf(
-				err, "unable to look up overlapping replica on %s", exReplica))
-		} else {
-			inactive := func(r *Replica) bool {
-				if r.RaftStatus() == nil {
-					return true
-				}
-				lease, pendingLease := r.GetLease()
-				now := s.Clock().Now()
-				return !r.IsLeaseValid(lease, now) &&
-					(pendingLease == nil || !r.IsLeaseValid(*pendingLease, now))
-			}
-
-			// If the existing range shows no signs of recent activity, give it a GC
-			// run.
-			if inactive(exReplica) {
-				if _, err := s.replicaGCQueue.Add(exReplica, replicaGCPriorityCandidate); err != nil {
-					log.Errorf(ctx, "%s: unable to add replica to GC queue: %s", exReplica, err)
-				} else {
-					msg += "; initiated GC:"
-				}
-			}
-		}
-		return nil, errors.Errorf("%s %v", msg, exReplica) // exReplica can be nil
-	}
-
-	placeholder := &ReplicaPlaceholder{
-		rangeDesc: *rangeDescriptor,
-	}
-	return placeholder, nil
 }
 
 func (s *Store) updateCapacityGauges() error {
